Move plugins code under src/worker
