+++ src/worker/plugins/loadPlugin.ts
@@ -1,9 +1,9 @@
 import * as fs from 'fs'
 import * as path from 'path'
 
-import { processError } from '../error'
-import { PluginConfig, PluginJsonConfig, PluginsServer } from '../types'
-import { getFileFromArchive, pluginDigest } from '../utils'
+import { processError } from '../../error'
+import { PluginConfig, PluginJsonConfig, PluginsServer } from '../../types'
+import { getFileFromArchive, pluginDigest } from '../../utils'
 
 export async function loadPlugin(server: PluginsServer, pluginConfig: PluginConfig): Promise<boolean> {
     const { plugin } = pluginConfig
